+++ src/main/java/com/cdweb/be/repository/ReviewRepository.java
@@ -14,28 +14,31 @@ import org.springframework.stereotype.Repository;
 public interface ReviewRepository extends JpaRepository<Review, Integer> {
 
   Page<Review> findByProductIdAndIsApproved(
-      Integer productId, Boolean isApproved, Pageable pageable);
+          Integer productId, Boolean isApproved, Pageable pageable);
 
   @Query(
-      "SELECT r FROM Review r WHERE r.product.id = :productId AND r.isApproved = true AND r.rating = :rating")
+          "SELECT r FROM Review r WHERE r.product.id = :productId AND r.isApproved = true AND r.rating = :rating")
   Page<Review> findByProductIdAndRating(
-      @Param("productId") Integer productId, @Param("rating") Integer rating, Pageable pageable);
+          @Param("productId") Integer productId, @Param("rating") Integer rating, Pageable pageable);
 
-  boolean existsByProductIdAndUserId(Integer productId, Integer userId);
+  // Đã sửa Integer -> Long userId
+  boolean existsByProductIdAndUserId(Integer productId, Long userId);
 
-  Optional<Review> findByIdAndUserId(Integer id, Integer userId);
+  // Đã sửa Integer -> Long userId
+  Optional<Review> findByIdAndUserId(Integer id, Long userId);
 
+  // Đã sửa Integer -> Long userId
   @Query("SELECT r FROM Review r WHERE r.user.id = :userId ORDER BY r.createdAt DESC")
-  Page<Review> findByUserId(@Param("userId") Integer userId, Pageable pageable);
+  Page<Review> findByUserId(@Param("userId") Long userId, Pageable pageable);
 
   @Query(
-      "SELECT AVG(r.rating) FROM Review r WHERE r.product.id = :productId AND r.isApproved = true")
+          "SELECT AVG(r.rating) FROM Review r WHERE r.product.id = :productId AND r.isApproved = true")
   Double findAverageRatingByProductId(@Param("productId") Integer productId);
 
   @Query("SELECT COUNT(r) FROM Review r WHERE r.product.id = :productId AND r.isApproved = true")
   Integer countApprovedByProductId(@Param("productId") Integer productId);
 
   @Query(
-      "SELECT r.rating, COUNT(r) FROM Review r WHERE r.product.id = :productId AND r.isApproved = true GROUP BY r.rating")
+          "SELECT r.rating, COUNT(r) FROM Review r WHERE r.product.id = :productId AND r.isApproved = true GROUP BY r.rating")
   List<Object[]> countByProductIdGroupByRating(@Param("productId") Integer productId);
 }
